Minor fixes
No errors in previous versions resulting from the problems fixed in this commit were actually experienced in the past.

It's possible that earlier versions would throw an error in the case where the destination account is commercial, since the endpoint for using refresh tokens was incorrect in this case. My guess is that no errors were thrown in testing because not enough documents were migrated in order for enough time to pass for using a refresh token to be necessary.

The following types of migrations were confirmed to work in this commit:

- commercial -> FedRAMP
- FedRAMP -> commercial
- commercial -> commercial

FedRAMP -> FedRAMP does not currently work because the application caches the email of whichever FedRAMP account is logged into first and attempts to use it to log into the second account, causing an error.